[rector] apply SetList::DEAD_CODE, better type checking (see msg)
This commit includes a variety of changes that improve the codebase's quality and consistency. The following changes were made:

- Replaced `==` with `===` for strict comparisons.
- Removed redundant `break` statements after `return` in switch-case structures.
- Fixed variable naming consistency and unnecessary variable assignments.
- Simplified conditional statements and removed unused code.
- Corrected minor typos and formatted code for better readability.
- Enhanced type safety by casting variables explicitly and adding type hints.
- Improved function and variable documentation with consistent comments.
- Refactored repeated logic to improve maintainability.
- Removed redundant checks for null values and unnecessary extractions.

+++ src/ChurchCRM/utils/InputUtils.php
@@ -55,12 +55,12 @@ class InputUtils
 
     public static function filterInt($sInput): int
     {
-        return (int) intval(trim($sInput));
+        return intval(trim($sInput));
     }
 
     public static function filterFloat($sInput): float
     {
-        return (float) floatval(trim($sInput));
+        return floatval(trim($sInput));
     }
 
     public static function filterDate($sInput): string
